# docs/swagger.yaml
basePath: /v1/api
definitions:
  go-app_internal_dto.ChangePasswordDto:
    properties:
      conf_password:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  go-app_internal_dto.ForgetPasswordDto:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  go-app_internal_dto.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  go-app_internal_dto.Logout:
    properties:
      refresh_token:
        type: string
    type: object
...
  go-app_internal_dto.RefreshTokenDto:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  go-app_internal_dto.RegisterDto:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  go-app_internal_dto.ResetPasswordDto:
    properties:
      confirm_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - confirm_password
    - new_password
    type: object
...
  go-app_internal_dto.VerifyOtpDto:
    properties:
      otp:
        type: string
    required:
    - otp
    type: object
  go-app_internal_dto.VerifyOtpResetDto:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: This is a sample server API for Go App.
  title: Go App API
  version: "1.0"
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Change Password Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.ChangePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Change password
      tags:
      - auth
  /auth/forget:
    post:
      consumes:
      - application/json
      description: Send OTP to email for password reset
      parameters:
      - description: Forget Password Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.ForgetPasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Forget password
      tags:
      - auth
  /auth/forget/verify:
    post:
      consumes:
      - application/json
      description: Verify OTP and return reset token
      parameters:
      - description: Verify OTP Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.VerifyOtpResetDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Verify forget password OTP
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return tokens
      parameters:
      - description: Login Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate user session/token
      parameters:
      - description: Logout Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.Logout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: User logout
      tags:
      - auth
...
  /auth/otp/verify:
    post:
      consumes:
      - application/json
      description: Validate OTP sent to user email
      parameters:
      - description: Verify OTP Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.VerifyOtpDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Validate OTP
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh Token Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.RefreshTokenDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.RegisterDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: User registration
      tags:
      - auth
  /auth/reset:
    post:
      consumes:
      - application/json
      description: Reset password using reset token
      parameters:
      - description: Reset Password Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.ResetPasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Reset password
      tags:
      - auth
...
swagger: "2.0"